refactor: use existing component for balance and fix select label style
