+++ src/components/Select/index.tsx
@@ -30,10 +30,10 @@ const SelectLabel = ({
 }: SelectLabelProps): JSX.Element => (
   <Listbox.Label
     className={clsx(
-      'block',
-      'text-sm',
-      'text-textSecondary',
-      'mb-1',
+      // TODO: placeholder color
+      { 'text-interlayTextSecondaryInLightMode': process.env.REACT_APP_RELAY_CHAIN_NAME === POLKADOT },
+      // TODO: placeholder color
+      { 'dark:text-kintsugiTextSecondaryInDarkMode': process.env.REACT_APP_RELAY_CHAIN_NAME === KUSAMA },
       className
     )}
     {...rest} />
