+++ src/pages/Staking/index.tsx
@@ -36,11 +36,11 @@ import Title from './Title';
 import BalancesUI from './BalancesUI';
 import WithdrawButton from './WithdrawButton';
 import ClaimRewardsButton from './ClaimRewardsButton';
-import AvailableBalanceUI from './AvailableBalanceUI';
 import InformationUI from './InformationUI';
 import LockTimeField from './LockTimeField';
 import MainContainer from 'parts/MainContainer';
 import Panel from 'components/Panel';
+import AvailableBalanceUI from 'components/AvailableBalanceUI';
 import TokenField from 'components/TokenField';
 import SubmitButton from 'components/SubmitButton';
 import ErrorFallback from 'components/ErrorFallback';
